backend/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai==1.0.0
flask-cors==4.0.0
email-validator==2.0.0
pytest==7.4.3
python-jose==3.3.0
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
